Add almost adequate non-java report generator #2687 Report generating process is rewritten and migrated to SAX Add bundled xml parser, bug #2612 Solve problems with java during building process, see bug #2607

+++ tools/CTesK/include/c_tracer/TraceFormatter.h
@@ -63,7 +63,7 @@ struct _TraceFormatter
   String * (*traceModelOperationResult)(void *state, int traceId, const char *type, const char *value);
   String * (*traceModelOperationChannel)(void *state, int traceId, const char *channel);
   String * (*traceModelOperationTimestamp)(void *state, int traceId, const char *timestamp);
-  String * (*traceOracleStart)(void *state, int traceId, const char *subsystem, const char *signature, unsigned ref);
+  String * (*traceOracleStart)(void *state, int traceId, const char *subsystem, const char *signature, long ref);
   String * (*tracePreconditionEnd)(void *state, int traceId);
   String * (*traceOracleEnd)(void *state, int traceId);
   String * (*traceModelOperationEnd)(void *state, int traceId);
